requirements.txt
# requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
aiosqlite>=0.19.0
certifi>=2024.2.2
python-dateutil>=2.8.2